Allow functions to be constant (MLDB-1984)
