+++ testing/testing.mk
@@ -438,3 +438,4 @@ $(eval $(call mldb_unit_test,MLDB-1950-crash-in-merge.py))
 $(eval $(call mldb_unit_test,MLDB-1921_merge_ds_strings.py))
 $(eval $(call mldb_unit_test,MLDB-1911_horizontal_agg_no_from.py))
 $(eval $(call mldb_unit_test,MLDB-1788_select_col_as_star_err.py))
+$(eval $(call mldb_unit_test,MLDB-1984-constant-functions.js))
